lib: better support for self-hosted instances
This change removes some instances of the hardcoded string
`https://sourcegraph.com`, and relies on the environment variable
`SRC_ENDPOINT`.

Additionally, it also makes some changes to how environment variables
are handled to make sure it matches the instructions in the src-cli
README, namely:

1. the environment variable for the token is `SRC_ACCESS_TOKEN`, not
   `SOURCEGRAPH_ACCESS_TOKEN`
2. the endpoint isn't guaranteed to have a trailing `/` (and the
   instructions in the README of src-cli don't mention it)

I don't feel strongly about those, but it's just something I noticed
while trying to follow the instructions in the README. If we don't want
to keep those, it may be worth it mentioning that a trailing slash is
required and that the environment variable for the access token is
different.

I haven't added any tests just yet, because I want to see what kind of
feedback this will get before investing in tests.

In general, I haven't been able to get it to work yet, but at this point
it's because some of the queries checked in here don't work with the
instance I have at work (probably a versioning thing). I'll keep digging
and thinking about how we can handle such differences.

Related to #11.