fix: prevent typewriter for answers we already saw
It you CodyToggle three times (on, off and on again), prompts sent by the
user are not rendered as typewrited, but the answers from Cody will be,
making the UI very annoying to deal with.

This is a crude fix I did in between things during my evening, should be
an okay bandage until we see if there's a better place to do this.

Basically, if a messeag has been rendered, it assumes it's been completed.